Split package into one deb per daemon
Restructured the packaging to create two debs rather than one.
Actually, three: a common, vxsnd, and vxrd.  This way, the user
installs only the one they want.  Now that each package has only one
daemon, the post install script starts the daemon as a service.  In
addition, monit is configured to automatically monitor the daemon,
restarting it if it stops or crashes.