Fast ageout when VTEP is removed from VXLAN
When the VXLAN config at a VTEP is changed such that a VXLAN interface
is removed, the RD now sends a refresh message with holdtime of 0 and
cleans up its peer list.  On receiving a refresh with holdtime 0,
vxsnd immediately deletes the FDB entry.  Before, cleanup occured when
the soft state expired.